+++ CloudHarvestCoreTasks/__register__.py
@@ -1,7 +1,6 @@
 from .chains import ReportTaskChain
 from .collectors import CacheDataCollector
 from .tasks import (
-    DelayTask,
     DummyTask,
     FileTask,
     HarvestRecordSetTask,